Fix: bug routingu dzienników
