+++ src/styles/JournalSection.scss
@@ -30,6 +30,8 @@
   max-width: 1350px;
   width: 100%;
 
+  margin: 0 auto;
+
   padding: 1em 0;
 }
